correct errors and add new todo list

+++ database/migrations/2020_03_16_195429_create_products_table.php
@@ -15,9 +15,9 @@ class CreateProductsTable extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('category_id')->nullable();
+            $table->unsignedBigInteger('category_id')->nullable();
             $table->string('product_name');
-            $table->float('product_price');
+            $table->float('product_price', 12, 2);
             $table->string('product_image');
             $table->timestamps();
         });
